Version V 1.0

+++ cerberus/core/views.py
@@ -137,7 +137,7 @@ def import_upload_data():
     from .tickets import parse_tickets, insert_tickets
     from .secretsdump import parse_secretsdump
 
-    
+
     # Importado de bibliotecas
     from io import BytesIO
 
@@ -197,7 +197,12 @@ def uploads_tickets_file():
     import os
     from cerberus.core.tickets import upload_ticket
     from cerberus import db2 as db
-    
+    from werkzeug.utils import secure_filename
+
+    def allowed_file(filename):
+        allowed_extensions = {'ccache', 'kirbi'}
+        return '.' in filename and filename.rsplit('.', 1)[1].lower() in allowed_extensions
+
     try:
         db.db_link(request.cookies.get('project'))
         db.db_connect()
@@ -211,17 +216,22 @@ def uploads_tickets_file():
             flash(u'No files selected', 'warning')
             return redirect(url_for('core.import_data'))
         for file in files:
-            file.save(f"{current_app.root_path}/static/uploads/tickets/{file.filename}")
-            #Ruta del ticket, nombre del archivo y session en la base de datos
-            upload_ticket(f"{current_app.root_path}/static/uploads/tickets/", file.filename ,db.session)
-            os.remove(f"{current_app.root_path}/static/uploads/tickets/{file.filename}")
-
+            filename = secure_filename(file.filename)
+            if(allowed_file(filename)):
+                file.save(f"{current_app.root_path}/static/uploads/tickets/{file.filename}")
+                #Ruta del ticket, nombre del archivo y session en la base de datos
+                upload_ticket(f"{current_app.root_path}/static/uploads/tickets/", file.filename ,db.session)
+                os.remove(f"{current_app.root_path}/static/uploads/tickets/{file.filename}")
+            else:
+                flash(u'Something went wrong: Wrong Extension only ccache and kirbi: ' + filename, 'danger')
+                print(f"\n\nERROR {filename}\n\n")
+                return redirect(url_for('core.import_data'))
         
         db.db_disconnect()
         flash(u'Tickets successfully uploaded', 'success')
         return redirect(url_for('core.import_data'))
     except Exception as e:
-        flash(u'Something went wrong', 'error')
+        flash(u'Something went wrong', 'danger')
         return redirect(url_for('core.import_data'))
 
 @core.route('/import/nmapxml', methods=['POST'])
@@ -232,13 +242,17 @@ def uploads_nmapxml_file():
     import os
     from cerberus.core.nmapxml import parseXML,uploadDataNmap
     from cerberus import db2 as db
-    
+    from werkzeug.utils import secure_filename
+    def allowed_file(filename):
+        allowed_extensions = {'xml'}
+        return '.' in filename and filename.rsplit('.', 1)[1].lower() in allowed_extensions
+
+
     try:
         db.db_link(request.cookies.get('project'))
         db.db_connect()
 
-        
-
+    
         files = request.files.getlist('dataFileNmaps')
 
         #if files are empty then raise error
@@ -246,12 +260,16 @@ def uploads_nmapxml_file():
             flash(u'No files selected', 'warning')
             return redirect(url_for('core.import_data'))
         for file in files:
-            file.save(f"{current_app.root_path}/static/.tmp/{file.filename}")
-            data = parseXML(f"{current_app.root_path}/static/.tmp/{file.filename}")
-            uploadDataNmap(data,db.session,file.filename)
-            os.remove(f"{current_app.root_path}/static/.tmp/{file.filename}")
             
-        
+            filename = secure_filename(file.filename)
+            if(allowed_file(filename)):
+                file.save(f"{current_app.root_path}/static/.tmp/{file.filename}")
+                data = parseXML(f"{current_app.root_path}/static/.tmp/{file.filename}")
+                uploadDataNmap(data,db.session,file.filename)
+                os.remove(f"{current_app.root_path}/static/.tmp/{file.filename}")
+            else:
+                flash(u'Something went wrong: Wrong Extension only XML:' + filename, 'danger')
+                return redirect(url_for('core.import_data'))
 
         db.db_disconnect()
         flash(u'Nmap successfully uploaded', 'success')
@@ -276,7 +294,12 @@ def import_ad_file():
     from io import BytesIO
     import subprocess
     from cerberus.static.external_modules.adexplorersnapshot.importDatFile import importDatFile
+    from werkzeug.utils import secure_filename
 
+    def allowed_file(filename):
+        allowed_extensions = {'zip', 'json', 'dat'}
+        return '.' in filename and filename.rsplit('.', 1)[1].lower() in allowed_extensions
+    
     async def parse_files(files):
         driver = database.init_driver(g.user["neo4j_url"], g.user["neo4j_user"], g.user["neo4j_password"])
         try:
@@ -312,36 +335,47 @@ def import_ad_file():
         return redirect(url_for('core.import_data'))
     
     if ".dat" in fileType:
-        upload_path = f"cerberus/static/.tmp/{files[0].filename}"
-        stream = BytesIO(files[0].read())
-        utils.saveFile(stream, upload_path)
-        # Parseamos el .dat a formato interpretable
-        zip_path = f"cerberus/static/.tmp/{files[0].filename.split('.')[0]}.zip"
-        importDatFile(upload_path, zip_path)
-        # Importamos el zip generado
-        try:
-            asyncio.run(parse_files([zip_path]))
-        except Exception as e:
+        filename = secure_filename(files[0].filename)
+        if(allowed_file(filename)):
+            upload_path = f"cerberus/static/.tmp/{filename}"
+            stream = BytesIO(files[0].read())
+            utils.saveFile(stream, upload_path)
+            # Parseamos el .dat a formato interpretable
+            zip_path = f"cerberus/static/.tmp/{files[0].filename.split('.')[0]}.zip"
+            importDatFile(upload_path, zip_path)
+            # Importamos el zip generado
+            try:
+                asyncio.run(parse_files([zip_path]))
+            except Exception as e:
+                try:
+                    subprocess.Popen(['rm', upload_path])
+                    subprocess.Popen(['rm', zip_path])
+                except:
+                    pass
+                print(e)
+                flash(u'Error loading file', 'danger')
+                return redirect(url_for('core.import_data'))
             try:
                 subprocess.Popen(['rm', upload_path])
                 subprocess.Popen(['rm', zip_path])
             except:
                 pass
-            print(e)
-            flash(u'Error loading file', 'danger')
+        else:     
+            flash(u'Something went wrong: Wrong Extension only .dat:' + filename, 'danger')
             return redirect(url_for('core.import_data'))
-        try:
-            subprocess.Popen(['rm', upload_path])
-            subprocess.Popen(['rm', zip_path])
-        except:
-            pass
     else:
         # Save files
         for i in files:
-            upload_path = f"cerberus/static/.tmp/{i.filename}"
-            stream = BytesIO(i.read())
-            utils.saveFile(stream, upload_path)
-            files_names.append(upload_path)
+            filename = secure_filename(i.filename)
+            if(allowed_file(filename)):
+                upload_path = f"cerberus/static/.tmp/{filename}"
+                stream = BytesIO(i.read())
+                utils.saveFile(stream, upload_path)
+                files_names.append(upload_path)
+            else:
+                flash(u'Something went wrong: Wrong Extension only json or zip:' + filename, 'danger')
+                return redirect(url_for('core.import_data'))
+
         # Parse files
         #try:
         asyncio.run(parse_files(files_names))
@@ -395,7 +429,7 @@ def hashcat_page():
         db.db_disconnect()
         return redirect(request.referrer)
     except:
-        flash(u'Error reading file', 'error')
+        flash(u'Error reading file', 'danger')
         # cerrar la conexión a la bbdd
         db.db_disconnect()
         return redirect(request.referrer)
@@ -454,7 +488,7 @@ def machines_dnsresolver():
         db.db_disconnect()
         return redirect(request.referrer)
     except:
-        flash(u'Error reading file', 'error')
+        flash(u'Error reading file', 'danger')
         # cerrar la conexión a la bbdd
         db.db_disconnect()
         return redirect(request.referrer)
